requirements.txt
Django
django-health-check
django-tastypie
django-nose